requirements.txt
requests
aiohttp
pyperclip
pycryptodome